work on stock modual and other changes

+++ app/Http/Livewire/Product/Detail.php
@@ -166,6 +166,7 @@ class Detail extends Component
         'product.sku' => '',
         'product.barcode' => '',
         'product.hscode' => '',
+        'product.stock' => '',
         'product.location' => '',
         'product.product_new' => '',
         'product.trackqtn' => '',
@@ -175,6 +176,7 @@ class Detail extends Component
         'Productvariant.*.sku' => 'required|string',
         'Productvariant.*.barcode' => 'required|string',
         'Productvariant.*.hscode' => 'required|string',
+        'Productvariant.*.stock' => 'required|string',
         'variantStock.*.stock' => 'required',
         'Productvariant.*.photo' => [],
         'att_price' => [],
@@ -354,11 +356,12 @@ class Detail extends Component
         $varition_arrray_crunch = $request['varition_arrray'];
         $price_arr = $request['att_price'];
         $price_selling_arr = $request['att_price_selling'];
-        $cost_arr = $request['att_cost'];
+        $stock_single_arr = $request['att_stock_qtn'];
+/*        $cost_arr = $request['att_cost'];
         $sku_arr = $request['att_sku'];
         $barcode_arr = $request['att_barcode'];
         $hscode_arr = $request['att_hscode'];
-        $country_arr = $request['att_country'];
+        $country_arr = $request['att_country'];*/
         $margin_arr = $request['margin_arry'];
         $profit_arr = $request['profit_arry'];
         $variations_arr = [];
@@ -373,11 +376,12 @@ class Detail extends Component
                 $variations['product_id'] = $request['product_id'];
                 $variations['price'] = $price_arr[$key];
                 $variations['selling_price'] = $price_selling_arr[$key];
-                $variations['cost'] = $cost_arr[$key];
+                $variations['stock'] = $stock_single_arr[$key];
+                /*$variations['cost'] = $cost_arr[$key];
                 $variations['sku'] = $sku_arr[$key];
                 $variations['barcode'] = $barcode_arr[$key];
                 $variations['hscode'] = $hscode_arr[$key];
-                $variations['country'] = $country_arr[$key];
+                $variations['country'] = $country_arr[$key];*/
                 $variations['margin'] = $margin_arr[$key];
                 $variations['profit'] = $profit_arr[$key];
                 // dd($explode_array);
@@ -535,6 +539,8 @@ class Detail extends Component
                     
                     'compare_price'    => $this->product['compare_price'],
                     
+                    'stock'            => $this->product['stock'],
+                    
                     'cost'             => $this->product['cost'],
                     
                     'margin'           => $this->product['margin'],
@@ -594,7 +600,8 @@ class Detail extends Component
                    'selling_price' => $this->Productvariant[$key]['selling_price'],
                    'sku' => $this->Productvariant[$key]['sku'],
                    'barcode' => $this->Productvariant[$key]['barcode'],
-                   'hscode' => $this->Productvariant[$key]['hscode']
+                   'hscode' => $this->Productvariant[$key]['hscode'],
+                   'stock' => $this->Productvariant[$key]['stock']
 
                 ]);
             } 
@@ -628,10 +635,11 @@ class Detail extends Component
         $deleteimg = "";
         foreach ($unlinkimg as $key => $delete) {
 
+            if (file_exists("app/public/{$delete->image}")){
             $image_path = storage_path("app/public/{$delete->image}");
-           $deleteimg = unlink($image_path);
 
-            
+            $deleteimg = unlink($image_path);
+            }
         }
